% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_relevant_functions.R
\name{find_identical_row_index}
\alias{find_identical_row_index}
\title{Find Index of an Identical Row in a Data Frame}
\usage{
find_identical_row_index(row, dataframe)
}
\arguments{
\item{row}{A data frame row or vector to search for}

\item{dataframe}{The data frame to search within}
}
\value{
An integer representing the index of the first identical row found,
or NA if no identical row exists in the data frame
}
\description{
Searches a data frame for a row that is identical to the provided row
and returns its index.
}
\details{
This function iterates through each row of the provided data frame and compares it
with the target row using the all.equal function. The comparison ignores attributes
to focus on the content of the data.

The function is particularly useful for finding specific mutations or cells in larger
data frames when the exact row index is not known but the content is.
}
\seealso{
\code{\link{sim_snv_nt_sc}}
}
