% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genome}
\alias{genome}
\title{Mini Genome Data}
\format{
A list containing two DNAStringSet objects for maternal and paternal haplotypes:
\describe{
\item{maternal}{DNAStringSet with three chromosomes (chr1, chr2, chr3)}
\item{paternal}{DNAStringSet with three chromosomes (chr1, chr2, chr3)}
}
}
\usage{
genome
}
\description{
A small simulated genome dataset for tutorial purposes, containing three chromosomes
of different lengths (100bp, 200bp, and 300bp).
}
\details{
Chromosome lengths:
\itemize{
\item chr1: 100 base pairs
\item chr2: 200 base pairs
\item chr3: 300 base pairs
}

Used in package tutorials to demonstrate basic functionality before moving to
whole-genome simulations.
}
\keyword{datasets}
