% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_relevant_functions.R
\name{get_clone_ancestors}
\alias{get_clone_ancestors}
\title{Get Ancestor Nodes in Phylogenetic Tree}
\usage{
get_clone_ancestors(tree, node)
}
\arguments{
\item{tree}{An igraph tree object representing the phylogenetic structure}

\item{node}{The target node for which to find ancestors}
}
\value{
A character vector containing the names of the ancestor nodes, ordered from
the root to the immediate parent of the target node.
}
\description{
This function retrieves the ancestors of a specified node (clone) in a tree structure.
It calculates the shortest path from the root node to the target node and returns the names of
all ancestor nodes (excluding the target node itself).
}
\details{
The function:
\enumerate{
\item Identifies the root node (assumes first vertex is root)
\item Finds shortest path from root to target node
\item Returns names of all nodes in the path
}
}
\note{
Assumes the tree is properly rooted with root node as first vertex
}
\examples{
\dontrun{
library(igraph)
# Create a simple tree
tree <- make_tree(3, 2)
V(tree)$name <- c("root", "A", "B")

# Get ancestors of node "B"
ancestors <- get_clone_ancestors(tree, "B")
# Returns: c("root", "B")
}

}
