% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_relevant_functions.R
\name{get_edges_between_clones}
\alias{get_edges_between_clones}
\title{Get Edge Names Between Two Nodes in Tree}
\usage{
get_edges_between_clones(tree, upper_node, lower_node)
}
\arguments{
\item{tree}{An igraph object representing the phylogenetic tree structure}

\item{upper_node}{Character. Name of the starting (ancestor) node}

\item{lower_node}{Character. Name of the ending (descendant) node}
}
\value{
A character vector containing edge names in the format "parent_child"
for all edges in the path from upper_node to lower_node
}
\description{
Finds and returns the names of all edges in the path between two nodes in a phylogenetic tree,
formatted as "parent_child" strings.
}
\details{
The function:
\enumerate{
\item Finds the shortest path between the two nodes
\item Identifies all edges along this path
\item Formats edge names as "parent_child"
}

Note: Assumes that the nodes are connected and upper_node is an ancestor of lower_node
in the tree structure.
}
\examples{
\dontrun{
# For a tree with path A -> B -> C
edges <- get_edges_between_clones(tree, "A", "C")
# Returns: c("A_B", "B_C")
}
}
